Crate harmony_build

This crate helps you compile the Harmony protocol into Rust code.

Note that you will need to include prost in your crate dependencies if you are compiling the protocols. hrpc is also needed if you are compiling client and server code.

Structs

Builder

Type containing configuration for building a Protocol.

Protocol

Contains paths for protocol includes and protos to compile.

Type Definitions

BoxError

A boxed object that implements std::error::Error.

Result

Shorthand for Result<T, BoxError>.